fixed auth服务循环依赖问题
Description:

The dependencies of some of the beans in the application context form a cycle:

   tokenController (field private com.xueyi.auth.service.SysLoginService com.xueyi.auth.controller.TokenController.sysLoginService)
      ↓
   sysLoginService (field private com.xueyi.system.api.feign.RemoteLogService com.xueyi.auth.service.SysLoginService.remoteLogService)
      ↓
   org.springframework.boot.autoconfigure.web.servlet.WebMvcAutoConfiguration$EnableWebMvcConfiguration
┌─────┐
|  com.alibaba.cloud.sentinel.SentinelWebAutoConfiguration (field private java.util.Optional com.alibaba.cloud.sentinel.SentinelWebAutoConfiguration.sentinelWebInterceptorOptional)
└─────┘

Action:

Relying upon circular references is discouraged and they are prohibited by default. Update your application to remove the dependency cycle between beans. As a last resort, it may be possible to break the cycle automatically by setting spring.main.allow-circular-references to true.